requirements.txt
Flask==3.1.0
Flask_HTTPAuth==4.8.0
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
Requests==2.32.3
